Add support for reading /etc/mime.types
There is now a rudimentary parser for /etc/mime.types style files. Also,
the STANDARD-MIME-TYPE class now also provides a FILE-EXTENSIONS slot,
which may be used to track frequently used filename extensions for MIME
types.

Note, that it is the application's responsibility to load the file
extensions via (say) LOAD-MIME-FILE-ASSOCIATIONS, and to use this information
appropriately. The library itself provides only the mechanism to do so,
but doesn't provide and file name extensions itself.